build: Allow being built if linux/securebits.h isn't available
For some reason the RHEL6 kernel-headers package doesn't have it.

Let's just follow for now the cargo culting of "define defines ourself
if not available" that various kernel-tied utilities have because
various buildsystems are too shitty to make it easy to install newer
kernel headers even if you're running an old kernel.